requirements.txt
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
Cython==0.29.23
cython-bbox==0.1.3
easydict==1.9
h5py==3.3.0
idna==2.10
jsonpatch==1.32
jsonpointer==2.1
kiwisolver==1.3.1
m3u8downloader==0.10.0
matplotlib==3.2.2
mkl-fft
mkl-random
mkl-service
munkres==1.1.4
natsort==7.0.1
numpy==1.19.1
olefile==0.46
opencv-python==4.5.2.54
Pillow==8.0.0
pip==21.1.2
protobuf==4.0.0rc2
pycocotools==2.0.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
scipy==1.5.2
setuptools==52.0.0.post20210125
six==1.15.0
tensorboardX==2.4
terminaltables==3.1.0
timm==0.1.20
torch==1.7.1
torchfile==0.1.0
torchvision
tornado==6.1
tqdm==4.61.1
typing-extensions==3.10.0.0
urllib3==1.26.6
visdom==0.1.8.9
websocket-client==1.1.0
wells==1.5.0
wheel==0.36.2
